+++ changelog.md
@@ -1,6 +1,7 @@
 # 0.0.7
 
 * Added `generatePortalsAutomatically` command and config to toggle automatic portal generation
+* Fixed desyncs when teleporting on a vehicle
 
 # 0.0.6